Added the ability to go threw doors for prisoners
Major change currntly in test. the Ability for prisoners to go threw doors with minimal performance loss compared to the lock mod.

The new way it work: it uses the labor area to determine what doors can a prisoner reach or not in Patch_InjectLockCheck (Thanks to locks for making this somewhat easy todo)